(~) ComponentSettingsDialogViewModel: extract S-matrix display refresh into partial (under 500-line limit)
The geometry-key wiring pushed the file to 504 lines (hard limit 500, grandfather
list full). Move RefreshEntries/RefreshEffectiveEntries into a .Display.cs partial.
